+++ escola/views_tarefa.py
@@ -1,20 +1,22 @@
 #  Developed by Vinicius José Fritzen
-#  Last Modified 25/04/19 13:44.
+#  Last Modified 25/04/19 23:11.
 #  Copyright (c) 2019  Vinicius José Fritzen and Albert Angel Lanzarini
 
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.views import redirect_to_login
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
+from rolepermissions.checkers import has_object_permission
 
 from escola.forms import TarefaForm, ComentarioTarefaForm
 from escola.models import Turma, Tarefa, TarefaComentario
 
 
-# FIXME: 25/04/2019 por wwwvi: Pedir permissão
 def add_tarefa(request, turma_pk):
-    # FIXME Adicionar permissões, a lista de permissões do grupo LIDER, VICELIDER e REGENTE da turma;
     turma = get_object_or_404(Turma, pk=turma_pk)
+    if not has_object_permission('add_tarefa', request.user, turma):
+        return redirect_to_login(request.get_full_path())
     if request.method == 'POST':
         form = TarefaForm(turma, request.POST)
         if form.is_valid():
@@ -22,6 +24,8 @@ def add_tarefa(request, turma_pk):
             tarefa.titulo = form.cleaned_data['titulo']
             tarefa.turma = turma
             tarefa.materia = form.cleaned_data['materia']
+            if not has_object_permission('add_tarefa_mat', request.user, form.cleaned_data['materia']):
+                return redirect_to_login(request.get_full_path())
             tarefa.tipo = form.cleaned_data['tipo']
             tarefa.descricao = form.cleaned_data['descricao']
             tarefa.deadline = form.cleaned_data['deadline']
@@ -29,9 +33,6 @@ def add_tarefa(request, turma_pk):
             seg = tarefa.get_seguidor_manager()
             seg.adicionar_seguidor(request.user)
             seg.adicionar_seguidor(tarefa.materia.professor.user)
-
-            # TODO: 25/04/2019 por wwwvi: Dar permissão para o criador
-
             return HttpResponseRedirect(reverse('escola:list-materias', args=[turma_pk]))
     else:
         form = TarefaForm(turma=turma)
@@ -59,6 +60,8 @@ def list_tarefa(request, turma_pk):
 def edit_tarefa(request, tarefa_pk):
     tarefa = get_object_or_404(Tarefa, pk=tarefa_pk)
     turma = tarefa.turma
+    if not has_object_permission('edit_tarefa', request.user, tarefa):
+        return redirect_to_login(request.get_full_path())
     if request.method == 'POST':
         form = TarefaForm(turma, request.POST)
         if form.is_valid():
@@ -80,9 +83,10 @@ def edit_tarefa(request, tarefa_pk):
     return render(request, 'escola/tarefas/formTarefa.html', context=context)
 
 
-# TODO: 25/04/2019 por wwwvi: Adicionar permissões
 def delete_tarefa(request, tarefa_pk):
     tarefa = get_object_or_404(Tarefa, pk=tarefa_pk)
+    if not has_object_permission('delete_tarefa', request.user, tarefa):
+        return redirect_to_login(request.get_full_path())
     tarefa.delete()
     return HttpResponseRedirect(reverse('escola:index'))
